Added check if contrib files exist
git-svn-id: https://euclid.esac.esa.int/svn/ESA/SOC/SOC-3-DEV/SOC-3-07-QLook/QPF/branches/dev.HMIalone@14454 463500fe-d6be-4184-9b20-47717e75ecac

+++ scripts/BuildQPF.sh
@@ -170,8 +170,11 @@ install_scpt () {
 install_contrib () {
     local fil=$1
     local tgtdir=$2
+    bn=$(basename "${CONTRIB_PATH}/${fil}")
     say "  - Installing file $fil"
-    perform cp "'${CONTRIB_PATH}/${fil}'" "'${WORK_AREA}/${tgtdir}'"
+    if [ ! -f "${WORK_AREA}/${tgtdir}/${bn}" ]; then
+	perform cp "'${CONTRIB_PATH}/${fil}'" "'${WORK_AREA}/${tgtdir}'"
+    fi
 }
 
 ###### Start
